Fixed uart disable to allow minor delay
This is an undocumented issue with seldom reporting in the community:
If the device goes to sleep righ after something has been sent over uart,
the AVR's uart buffers get messed up and remain so even when the uart is
powered off before sleep and then on again after wakeup. The datasheet
says that the TXC register must be checked, but even when it is the
buffers still get messed up. The only viable solution was to add a
25 micro-second delay before disabling the uart interface, to allow
for the buffers to be cheared. This seems to work and since the time delay
is pretty small, I don't think I'll try to fix this any further.