Fixing TWI STOP bug
Apparently, when the AVR sets the stop condition on the TWI lines,
you don't have to wait for the TWINT flag. If you do, you'll never
get it, which makes the entire program hang. The stop condition
doesn't affect any buffers, so there's no need to wait. Just set the
stop condition and leave.